Restructured `splicing` module.
